# Convert from AngleAxis straight to RotMatrix using Rodrigues' rotation formula