Also make conversion from `RodriguesVec` to `RotMatrix` go through `AngleAxis` instead of `Quat`.

This is significantly faster.

### `AngleAxis` --> `RotMatrix`
```julia
@benchmark RotMatrix(aa) setup = aa = rand(AngleAxis)
```

Before:
```
BenchmarkTools.Trial: 
  memory estimate:  0.00 bytes
  allocs estimate:  0
  --------------
  minimum time:     42.916 ns (0.00% GC)
  median time:      57.375 ns (0.00% GC)
  mean time:        56.766 ns (0.00% GC)
  maximum time:     115.718 ns (0.00% GC)
```

After:
```
BenchmarkTools.Trial: 
  memory estimate:  0.00 bytes
  allocs estimate:  0
  --------------
  minimum time:     22.549 ns (0.00% GC)
  median time:      32.279 ns (0.00% GC)
  mean time:        32.430 ns (0.00% GC)
  maximum time:     85.305 ns (0.00% GC)
```


### `RodriguesVec` --> `RotMatrix`
```julia
@benchmark RotMatrix(vec) setup = vec = rand(RodriguesVec)
```

Before:
```
BenchmarkTools.Trial: 
  memory estimate:  0.00 bytes
  allocs estimate:  0
  --------------
  minimum time:     55.586 ns (0.00% GC)
  median time:      72.386 ns (0.00% GC)
  mean time:        71.351 ns (0.00% GC)
  maximum time:     107.514 ns (0.00% GC)
```

After:
```
BenchmarkTools.Trial: 
  memory estimate:  0.00 bytes
  allocs estimate:  0
  --------------
  minimum time:     36.108 ns (0.00% GC)
  median time:      47.739 ns (0.00% GC)
  mean time:        47.565 ns (0.00% GC)
  maximum time:     93.466 ns (0.00% GC)
```